refactor(brutus): extract protocol mapping to mapping.go
Move FingerprintxResult type and mapServiceToProtocol function.
Part of main.go refactoring (4/9).

Related: C1/C2 (split 1340-line main.go)
